requirements.txt
Flask==2.0.1
Flask_RESTful==0.3.9
Flask_SQLAlchemy==2.5.1
gunicorn
Flask-HTTPAuth==4.4.0
Flask-JWT==0.3.2